wifi: don't cast Unix time to uint32 for nm_ap_set_last_seen()
last_seen of NMAccessPoint is ever since of type glong. Casting
it to guint32 is wrong. Ok, it is not a real problem, because
we don't ever expect negative timestamps (pre 1970) or timestamps
larger then uint32 (year 2106). But let's just do it right.

+++ src/nm-device-wifi.c
@@ -2118,7 +2118,7 @@ supplicant_iface_bss_updated_cb (NMSupplicantInterface *iface,
 	/* Update the AP's last-seen property */
 	ap = get_ap_by_supplicant_path (self, object_path);
 	if (ap)
-		nm_ap_set_last_seen (ap, (guint32) time (NULL));
+		nm_ap_set_last_seen (ap, (glong) time (NULL));
 
 	/* Remove outdated access points */
 	schedule_scanlist_cull (self);
